Update to Cubism 4 SDK for Web R3

+++ src/motion/cubismmotionjson.ts
@@ -345,7 +345,7 @@ export class CubismMotionJson {
       .getValueByString(UserData)
       .getValueByIndex(userDataIndex)
       .getValueByString(Time)
-      .toInt();
+      .toFloat();
   }
 
   /**
